requirements.txt
Django==1.11.20
pytest-cov==2.5.1
mypy==0.600
psycopg2-binary==2.7.4
djangorestframework==3.7.3
jsonschema==2.6.0
numpy==1.14.3
pytest==3.5.1
pytest-django==3.2.1
google-api-python-client==1.6.5
google-auth-httplib2==0.0.3
google-auth==1.4.1
google-cloud-pubsub==0.32.1
cryptography==2.3
pyjwt==1.6.0
paho-mqtt==1.3.1
pyudev==0.21.0
whitenoise==4.1
django-bootstrap-static==4.0.0
django-tempus-dominus==5.0.1.2
django-jquery==3.1.0
pyftdi==0.29.2
pygame==1.9.4
